Divide the class into "output in column format" and "detailed output in row format"
- 「行形式でのファイル詳細出力(FileDetailRow)」と「列形式でのファイル名出力(FileNameColumn)」でクラスを分割
- FileDisplayクラス名をFileDisplayPreparationクラスに変更
- FileDisplayPreparationクラスのdisplay_filesメソッド名をdisplay_files_based_on_formatに変更
